Avoid go_repository restarts
See https://github.com/bazelbuild/bazel-gazelle/issues/1175 and https://github.com/bazelbuild/bazel-gazelle/pull/1206.

The pull request was conservative and only fixed restarts for repositories where it could be statically determined that `@bazel_gazelle_go_repository_tools//:bin/gazelle` was needed. It did not avoid restarts in the case where `go_repository` needs to fetch and examine the repo to determine if it requires `@bazel_gazelle_go_repository_tools`. But that's the common case for `go_repository`, so this patches the implementation to unconditionally require `@bazel_gazelle_go_repository_tools`.